I moved the Login request (and requisite JWT decoding) for initial user data into the generic-redirect component from the dashboard component and deleted it from the dashboard component
