+++ src/components/dashboard/Dashboard.js
@@ -10,25 +10,6 @@ import { axiosWithAuth } from "../../utilities/axiosWithAuth";
 import { print } from "graphql";
 
 const Dashboard = () => {
-  useEffect(() => {
-    if (ls.get("access_token")) {
-      const token = ls.get("access_token");
-      const decodedToken = jwt(token).sub;
-      axiosWithAuth()({
-        url: `${process.env.REACT_APP_BASE_URL}/graphql`,
-        method: "post",
-        data: {
-          query: print(USER_BY_EMAIL),
-          variables: { input: { email: decodedToken } },
-        },
-      }).then((res) => {
-        sessionStorage.setItem(
-          "user",
-          JSON.stringify(res.data.data.getUserByEmail)
-        );
-      });
-    }
-  }, []);
 
   return (
     <div className="dashboard-container">
